long -> long long型 に変更

+++ ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: rnishimo <rnishimo@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/13 13:03:01 by rnishimo          #+#    #+#             */
-/*   Updated: 2021/11/01 20:21:13 by rnishimo         ###   ########.fr       */
+/*   Updated: 2021/11/13 01:09:18 by rnishimo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	ft_putnbr_fd_recursive(long n, int fd)
+static void	ft_putnbr_fd_recursive(long long n, int fd)
 {
 	if (n < 0)
 	{
@@ -26,5 +26,5 @@ static void	ft_putnbr_fd_recursive(long n, int fd)
 
 void	ft_putnbr_fd(int n, int fd)
 {
-	ft_putnbr_fd_recursive((long)n, fd);
+	ft_putnbr_fd_recursive((long long)n, fd);
 }
